+++ NEWS.rst
@@ -7,6 +7,7 @@ of every change, see the Git log.
 Latest
 ------
 * Minor: Added mkspecs cxx_apple_llvm51_x86/64 for Apple LLVM 5.1 compiler.
+* Minor: Add cxx_ios50_apple_llvm51_armv7 makespec.
 
 2.26.0
 ------